fix(esp-idf): resolve remaining m5stick-c build errors
- Fix include paths for hal_system.h and hal_uart.h
- Migrate legacy timer driver to ESP-IDF 6.x GPTimer API
- Add missing FreeRTOS headers for portMUX_TYPE
- Fix format specifiers for uint32_t on xtensa
- Fix LovyanGFX setLight API usage (panel-level)
- Fix const qualifier mismatch in task stack usage
- Fix VFS child_count type overflow (uint8_t -> uint16_t)
- Fix missing stdio.h and popup struct initializers
- Remove static from g_canvas for cross-file linkage

+++ src/kernel/kern_shell.c
@@ -284,7 +284,7 @@ void shell_complete_path(kern_fd_t tty, char *line, size_t *pos,
     const char *matches[16];
     int match_count = 0;
 
-    for (uint8_t i = 0; i < dir->child_count && match_count < 16; i++) {
+    for (uint16_t i = 0; i < dir->child_count && match_count < 16; i++) {
         kern_dentry_t *child = dir->children[i];
         if (child == NULL) continue;
         if (dir_only && !shell_is_dir(child)) continue;
@@ -316,7 +316,7 @@ void shell_complete_path(kern_fd_t tty, char *line, size_t *pos,
 
         /* 找到匹配的 dentry 以判断类型 */
         kern_dentry_t *matched_d = NULL;
-        for (uint8_t i = 0; i < dir->child_count; i++) {
+        for (uint16_t i = 0; i < dir->child_count; i++) {
             if (dir->children[i] != NULL
                 && strcmp(dir->children[i]->name, match) == 0) {
                 matched_d = dir->children[i];
